requirements.txt
Django==1.8.2
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
Pillow==2.8.2
sqlparse==0.1.19
